Set the x_rh_identity if a tenant is passed in
If there is an external tenant passed in the payload then set the
x_rh_identity header for the topological_inventory_api_client.
